refactor(gleif-mcp-server): extract registration/schema wiring from handlers.go to drop under 500 LOC (deslop #39)
Move the Registry type, tool-registration machinery (NewRegistry,
buildHandlerMap, RegisterAll, registerTool, lookupHandler), the
input-schema builders (buildInputSchema/buildPropertySchema and the
add*Bounds helpers), and the panic-recovery wrapHandler into a new
sibling registry.go in package tools. Pure move, no symbol/signature/
logic changes. handlers.go 643->479; registry.go 173.
